Replace faucet with tap-min because of security issues
minimist < 0.2.1 is affected by a prototype pollution security vulnerability ([CVE-2020-7598](https://github.com/advisories/GHSA-vh95-rmgr-6w4m)).

Because faucet requires the fixed version of minimist (0.0.5), it is impossible to fix this issue.
faucet was last updated 8 years agon and seems dead; it is highly unlikely that the upstream is going to fix this issue.

This PR replaces faucet with tap-min.